feat: 게시글 페이지를 클라이언트 사이드 렌더링으로 변경
